requirements.txt
fastapi
uvicorn[standard]
google-generativeai
sentence-transformers
faiss-cpu
pandas
numpy
python-dotenv